refactor(backups): move Runner, VmBackup,  writers and specific method to a private folder
