+++ @xen-orchestra/backups/_runners/_VmBackup.js
@@ -14,15 +14,15 @@ const { defer } = require('golike-defer')
 const { formatDateTime } = require('@xen-orchestra/xapi')
 const { pipeline } = require('node:stream')
 
-const { IncrementalRemoteWriter } = require('./writers/IncrementalRemoteWriter.js')
-const { IncrementalXapiWriter } = require('./writers/IncrementalXapiWriter.js')
-const { exportIncrementalVm } = require('./_incrementalVm.js')
+const { IncrementalRemoteWriter } = require('./_writers/IncrementalRemoteWriter.js')
+const { IncrementalXapiWriter } = require('./_writers/IncrementalXapiWriter.js')
+const { exportIncrementalVm } = require('../_incrementalVm.js')
 const { forkStreamUnpipe } = require('./_forkStreamUnpipe.js')
-const { FullRemoteWriter } = require('./writers/FullRemoteWriter.js')
-const { FullXapiWriter } = require('./writers/FullXapiWriter.js')
-const { getOldEntries } = require('./_getOldEntries.js')
-const { Task } = require('./Task.js')
-const { watchStreamSize } = require('./_watchStreamSize.js')
+const { FullRemoteWriter } = require('./_writers/FullRemoteWriter.js')
+const { FullXapiWriter } = require('./_writers/FullXapiWriter.js')
+const { getOldEntries } = require('../_getOldEntries.js')
+const { Task } = require('../Task.js')
+const { watchStreamSize } = require('../_watchStreamSize.js')
 
 const { debug, warn } = createLogger('xo:backups:VmBackup')
