+++ @xen-orchestra/backups/_runners/_createStreamThrottle.js
