+++ @xen-orchestra/backups/_runners/_getAdaptersByRemote.js
