+++ @xen-orchestra/backups/_runners/_runTask.js
@@ -1,5 +1,5 @@
 'use strict'
-const { Task } = require('./Task.js')
+const { Task } = require('../Task.js')
 const noop = Function.prototype
 const runTask = (...args) => Task.run(...args).catch(noop) // errors are handled by logs
