+++ @xen-orchestra/backups/_runners/_writers/IncrementalRemoteWriter.js
@@ -11,9 +11,9 @@ const { decorateClass } = require('@vates/decorate-with')
 const { defer } = require('golike-defer')
 const { dirname } = require('path')
 
-const { formatFilenameDate } = require('../_filenameDate.js')
-const { getOldEntries } = require('../_getOldEntries.js')
-const { Task } = require('../Task.js')
+const { formatFilenameDate } = require('../../_filenameDate.js')
+const { getOldEntries } = require('../../_getOldEntries.js')
+const { Task } = require('../../Task.js')
 
 const { MixinRemoteWriter } = require('./_MixinRemoteWriter.js')
 const { AbstractIncrementalWriter } = require('./_AbstractIncrementalWriter.js')
