+++ @xen-orchestra/backups/_runners/_writers/IncrementalXapiWriter.js
@@ -4,10 +4,10 @@ const { asyncMap, asyncMapSettled } = require('@xen-orchestra/async-map')
 const ignoreErrors = require('promise-toolbox/ignoreErrors')
 const { formatDateTime } = require('@xen-orchestra/xapi')
 
-const { formatFilenameDate } = require('../_filenameDate.js')
-const { getOldEntries } = require('../_getOldEntries.js')
-const { importIncrementalVm, TAG_COPY_SRC } = require('../_incrementalVm.js')
-const { Task } = require('../Task.js')
+const { formatFilenameDate } = require('../../_filenameDate.js')
+const { getOldEntries } = require('../../_getOldEntries.js')
+const { importIncrementalVm, TAG_COPY_SRC } = require('../../_incrementalVm.js')
+const { Task } = require('../../Task.js')
 
 const { AbstractIncrementalWriter } = require('./_AbstractIncrementalWriter.js')
 const { MixinXapiWriter } = require('./_MixinXapiWriter.js')
