+++ @xen-orchestra/backups/_runners/_writers/_AbstractFullWriter.js
