+++ @xen-orchestra/backups/_runners/_writers/_AbstractIncrementalWriter.js
