+++ @xen-orchestra/backups/_runners/_writers/_AbstractWriter.js
