+++ @xen-orchestra/backups/_runners/_writers/_checkVhd.js
